+++ src/graphql/resolvers/expandable/updateOne.ts
@@ -20,9 +20,8 @@ const updateOneExpandable: T.Resolver<
     data: {
       name: data.name || undefined,
       key: data.key || undefined,
-      age: data.age || undefined,
       address: data.address || undefined,
-      description: data.description || undefined,
+      articles: data.articles ? data.articles.join(',') : undefined,
       updated: new Date(),
     },
   });
@@ -30,9 +29,8 @@ const updateOneExpandable: T.Resolver<
     id: result.id,
     name: result.name,
     key: result.key,
-    age: result.age,
+    articles: result.articles.split(',') || [],
     address: result.address,
-    description: result.description,
     created: result.created?.toISOString() || '',
     updated: result.updated?.toISOString() || '',
   };